+++ toqito/channel_ops/__init__.py
@@ -1,4 +1,5 @@
 """A number of operations on quantum channels."""
 from toqito.channel_ops.apply_map import apply_map
 from toqito.channel_ops.partial_map import partial_map
+from toqito.channel_ops.choi_to_kraus import choi_to_kraus
 from toqito.channel_ops.kraus_to_choi import kraus_to_choi